+++ python/GafferUI/__init__.py
@@ -272,7 +272,8 @@ from .TweakPlugValueWidget import TweakPlugValueWidget
 from .PlugPopup import PlugPopup
 from .OptionalValuePlugValueWidget import OptionalValuePlugValueWidget
 from .TogglePlugValueWidget import TogglePlugValueWidget
-from .PlugCreationWidget import PlugCreationWidget
+from .PlugCreationWidget import PlugCreationWidget, PlugCreationGadget
+from . import _PlugVisibilityGadget
 
 # and then specific node uis