+++ startup/GafferCyclesUI/cyclesExamples.py
@@ -86,3 +86,27 @@ GafferUI.Examples.registerExample(
 		GafferCycles.CyclesOptions
 	]
 )
+
+GafferUI.Examples.registerExample(
+	"Scene Processing/Shader Tweaks",
+	"$GAFFER_ROOT/resources/examples/sceneProcessing/shaderTweaks.gfr",
+	description = """
+	Demonstrates using ShaderTweaks for modifying shader networks.
+	""",
+	notableNodes = [
+		GafferScene.ShaderTweaks,
+		GafferScene.ShaderTweakProxy
+	]
+)
+
+GafferUI.Examples.registerExample(
+	"Scene Processing/Statistics",
+	"$GAFFER_ROOT/resources/examples/sceneProcessing/statistics.gfr",
+	description = "Demonstrates how to use the SceneStats node to add a statistics overlay to rendered images.",
+	notableNodes = [
+		GafferScene.PrimitiveQuery,
+		GafferScene.SceneStats,
+		GafferImage.ImageMetadata,
+		GafferImage.MetadataOverlay,
+	]
+)